Refactor state machine for unit testing
